chore: Update files structure
